+++ metadata.rb
@@ -1,10 +1,10 @@
-name             "sensu-citadel"
+name             "sensu"
 maintainer       "Sonian, Inc."
 maintainer_email "chefs@sonian.net"
 license          "Apache 2.0"
 description      "Installs/Configures Sensu"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "2.7.0"
+version          "2.10.99"
 
 # available @ http://supermarket.chef.io/cookbooks/apt
 depends "apt"