+++ providers/base_config.rb
@@ -1,12 +1,23 @@
 action :create do
   definitions = Sensu::Helpers.select_attributes(
-    node.sensu,
-    %w[rabbitmq redis api]
+    node["sensu"],
+    %w[transport rabbitmq redis api]
   )
 
-  config = JSON.parse(citadel["#{node.sensu.citadel.root}/config.json"])
-  unless config.empty?
-    definitions = Chef::Mixin::DeepMerge.merge(definitions, config)
+  if node["sensu"]["citadel"]["root"]
+    config = JSON.parse(citadel["#{node.sensu.citadel.root}/config.json"])
+    unless config.empty?
+      definitions = Chef::Mixin::DeepMerge.merge(definitions, config)
+    end
+  else
+    data_bag_name = node["sensu"]["data_bag"]["name"]
+    config_item = node["sensu"]["data_bag"]["config_item"]
+
+    config = Sensu::Helpers.data_bag_item(config_item, true, data_bag_name)
+
+    if config
+      definitions = Chef::Mixin::DeepMerge.merge(definitions, config.to_hash)
+    end
   end
 
   service_config = {}
@@ -16,11 +27,22 @@ action :create do
     api
     server
   ].each do |service|
-    next unless node.recipe?("sensu::#{service}_service")
+    unless node.recipe?("sensu::#{service}_service") ||
+        node.recipe?("sensu::enterprise_service")
+      next
+    end
 
-    service_config_item = JSON.parse(citadel["#{node.sensu.citadel.root}/#{service}_config.json"])
-    unless service_config_item.empty?
-      service_config = Chef::Mixin::DeepMerge.merge(service_config, service_config_item.to_hash)
+    if node["sensu"]["citadel"]["root"]
+      service_config_item = JSON.parse(citadel["#{node.sensu.citadel.root}/#{service}_config.json"])
+      unless service_config_item.empty?
+        service_config = Chef::Mixin::DeepMerge.merge(service_config, service_config_item.to_hash)
+      end
+    else
+      service_data_bag_item = Sensu::Helpers.data_bag_item(service, true, data_bag_name)
+
+      if service_data_bag_item
+        service_config = Chef::Mixin::DeepMerge.merge(service_config, service_data_bag_item.to_hash)
+      end
     end
   end
 
@@ -28,7 +50,7 @@ action :create do
     definitions = Chef::Mixin::DeepMerge.merge(definitions, service_config)
   end
 
-  f = sensu_json_file ::File.join(node.sensu.directory, "config.json") do
+  f = sensu_json_file ::File.join(node["sensu"]["directory"], "config.json") do
     content Sensu::Helpers.sanitize(definitions)
   end